+++ hbw/app/javascript/packs/hbw/components/helpers/index.js
@@ -2,6 +2,7 @@ import withCallbacks from './callbacks';
 import withConditions from './conditions';
 import withTasks from './tasks';
 import withSelect from './select';
+import withValidations from './validations';
 import { compose } from './utils';
 
 export {
@@ -9,5 +10,6 @@ export {
   withConditions,
   withTasks,
   withSelect,
+  withValidations,
   compose
 };