+++ hbw/app/javascript/packs/hbw/components/helpers/validations.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default WrappedComponent => (props) => {
+  const isFilled = value => (value !== null && value !== undefined && `${value}`.length > 0);
+
+  const isValid = value => props.params.nullable || isFilled(value);
+
+  return <WrappedComponent isValid={isValid}
+                           {...props} />;
+};